requirements.txt
Flask~=3.0.0
Flask-Login
Flask-WTF
Flask-SQLAlchemy
gunicorn
